Implement a monitoring of the size of responses for the given requests
Implement an additional Collector and Comparator that will allow one to sum the transferred data of the responses for the given/filtered requests. Added a simple report that shows the information about the overall amount of the transferred data and the table of ordered requests (desc) and their attributes (URL as a link, Size in kilobytes)

https://github.com/Cognifide/aet/issues/501

I hereby agree to the terms of the AET Contributor License Agreement.